% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chaburankingsmetadata}
\alias{chaburankingsmetadata}
\title{chaburankingsmetadata}
\format{
A data frame with 60 rows and 9 variables:
\describe{
\item{\code{date}}{double COLUMN_DESCRIPTION}
\item{\code{time}}{double COLUMN_DESCRIPTION}
\item{\code{id}}{double COLUMN_DESCRIPTION}
\item{\code{sex}}{character COLUMN_DESCRIPTION}
\item{\code{age_category}}{character COLUMN_DESCRIPTION}
\item{\code{age_estimated}}{double COLUMN_DESCRIPTION}
\item{\code{height}}{double COLUMN_DESCRIPTION}
\item{\code{grip_RH}}{double COLUMN_DESCRIPTION}
\item{\code{grip_LH}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://url.com/}
}
\usage{
chaburankingsmetadata
}
\description{
Participant metadata
}
\keyword{datasets}
